Don't offer a code action for a diagnostic with no diagnostic "code" (#100)
* Don't offer a code action for a diagnostic with no diagnostic "code"

Per the lsp spec, a diagnostic code, if one exists, should be a number
or a string.  At least one error message (the help message for missing
imports) doesn't include a code, so the parse_line function sets it to
"None".

When this happens, don't offer a code action for that diagnostic

* only include valid codes in diag

---------

Co-authored-by: Richard Kellnberger <git@richardk2n.de>